@@ -1,6 +1,6 @@
 {
   "name": "strapi-plugin-config-sync",
-  "version": "1.1.2",
+  "version": "2.0.0-beta.4",
   "description": "Migrate your config data across environments using the CLI or Strapi admin panel.",
   "strapi": {
     "displayName": "Config Sync",
@@ -14,12 +14,15 @@
     "config-sync": "./bin/config-sync"
   },
   "scripts": {
+    "develop": "nodemon -e js,jsx --ignore playground --exec \"yalc publish && yalc push\"",
+    "build": "yalc push --publish",
     "eslint": "eslint --max-warnings=0 './**/*.{js,jsx}'",
     "eslint:fix": "eslint --fix './**/*.{js,jsx}'",
     "test:unit": "jest --verbose",
-    "test:integration": "cd playground && node_modules/.bin/jest --verbose",
-    "plugin:install": "yarn install && rm -rf node_modules/@strapi/helper-plugin",
-    "playground:install": "cd playground && yarn install",
+    "test:integration": "cd playground && node_modules/.bin/jest --verbose  --forceExit --detectOpenHandles",
+    "playground:install": "yarn playground:yalc-add-link && cd playground && yarn install",
+    "playground:yalc-add": "cd playground && yalc add strapi-plugin-config-sync",
+    "playground:yalc-add-link": "cd playground && yalc add --link strapi-plugin-config-sync",
     "playground:build": "cd playground && yarn build",
     "playground:develop": "cd playground && yarn develop"
   },
@@ -31,19 +34,20 @@
     "git-diff": "^2.0.6",
     "immutable": "^3.8.2",
     "inquirer": "^8.2.0",
-    "react-diff-viewer-continued": "^3.2.3",
+    "react-diff-viewer-continued": "3.2.6",
+    "react-intl": "6.6.2",
     "redux-immutable": "^4.0.0",
     "redux-thunk": "^2.3.0"
   },
   "author": {
     "name": "Boaz Poolman",
-    "email": "info@boazpoolman.nl",
+    "email": "boaz@pluginpal.io",
     "url": "https://github.com/boazpoolman"
   },
   "maintainers": [
     {
       "name": "Boaz Poolman",
-      "email": "info@boazpoolman.nl",
+      "email": "boaz@pluginpal.io",
       "url": "https://github.com/boazpoolman"
     }
   ],
@@ -55,13 +59,13 @@
     "strapi-server.js"
   ],
   "peerDependencies": {
-    "@strapi/strapi": "^4.0.0"
+    "@strapi/strapi": "^5.0.0"
   },
   "devDependencies": {
-    "@strapi/design-system": "^1.3.1",
-    "@strapi/helper-plugin": "^4.5.5",
-    "@strapi/icons": "^1.3.1",
-    "@strapi/utils": "^4.5.2",
+    "@strapi/design-system": "2.0.0-rc.11",
+    "@strapi/icons": "2.0.0-rc.11",
+    "@strapi/strapi": "5.0.4",
+    "@strapi/utils": "5.0.4",
     "babel-eslint": "9.0.0",
     "eslint": "^7.32.0",
     "eslint-config-airbnb": "^18.2.1",
@@ -74,22 +78,23 @@
     "eslint-plugin-jsx-a11y": "^6.4.1",
     "eslint-plugin-react": "^7.21.5",
     "eslint-plugin-react-hooks": "^2.3.0",
-    "jest": "^29.3.1",
+    "jest": "^29.7.0",
     "jest-cli": "^29.3.1",
     "jest-styled-components": "^7.0.2",
     "lodash": "^4.17.11",
+    "nodemon": "^3.1.7",
     "react": "^17.0.2",
-    "react-intl": "^5.20.12",
     "react-redux": "^7.2.2",
     "redux": "^4.0.5",
-    "styled-components": "^5.2.3"
+    "styled-components": "^5.2.3",
+    "yalc": "^1.0.0-pre.53"
   },
   "bugs": {
-    "url": "https://github.com/boazpoolman/strapi-plugin-config-sync/issues"
+    "url": "https://github.com/pluginpal/strapi-plugin-config-sync/issues"
   },
-  "homepage": "https://github.com/boazpoolman/strapi-plugin-config-sync#readme",
+  "homepage": "https://www.pluginpal.io/plugin/config-sync",
   "engines": {
-    "node": ">=10.0.0",
+    "node": ">=18.0.0",
     "npm": ">=6.0.0"
   },
   "license": "MIT",
