@@ -1,8 +1,8 @@
-const pluginPkg = require('../../../package.json');
+import pluginPkg from '../../../package.json';
 
 const pluginId = pluginPkg.name.replace(
   /^strapi-plugin-/i,
   '',
 );
 
-module.exports = pluginId;
+export default pluginId;
