@@ -1,7 +1,7 @@
 import React from 'react';
-import { Tr, Td, BaseCheckbox } from '@strapi/design-system';
+import { Tr, Td, Checkbox, Typography } from '@strapi/design-system';
 
-const CustomRow = ({ row, checked, updateValue }) => {
+const CustomRow = ({ row, checked, updateValue, ...props }) => {
   const { configName, configType, state, onClick } = row;
 
   const stateStyle = (stateStr) => {
@@ -34,6 +34,7 @@ const CustomRow = ({ row, checked, updateValue }) => {
 
   return (
     <Tr
+      {...props}
       onClick={(e) => {
         if (e.target.type !== 'checkbox') {
           onClick(configType, configName);
@@ -42,20 +43,20 @@ const CustomRow = ({ row, checked, updateValue }) => {
       style={{ cursor: 'pointer' }}
     >
       <Td>
-        <BaseCheckbox
+        <Checkbox
           aria-label={`Select ${configName}`}
-          value={checked}
-          onValueChange={updateValue}
+          checked={checked}
+          onCheckedChange={updateValue}
         />
       </Td>
-      <Td>
-        <p>{configName}</p>
+      <Td onClick={(e) => props.onClick(e)}>
+        <Typography variant="omega">{configName}</Typography>
       </Td>
-      <Td>
-        <p>{configType}</p>
+      <Td onClick={(e) => props.onClick(e)}>
+        <Typography variant="omega">{configType}</Typography>
       </Td>
-      <Td>
-        <p style={stateStyle(state)}>{state}</p>
+      <Td onClick={(e) => props.onClick(e)}>
+        <Typography variant="omega" style={stateStyle(state)}>{state}</Typography>
       </Td>
     </Tr>
   );
