@@ -7,14 +7,15 @@
 import React, { memo } from 'react';
 import { useIntl } from 'react-intl';
 
-import { HeaderLayout, Box } from '@strapi/design-system';
+import { Layouts } from '@strapi/admin/strapi-admin';
+import { Box } from '@strapi/design-system';
 
 const HeaderComponent = () => {
   const { formatMessage } = useIntl();
 
   return (
     <Box background="neutral100">
-      <HeaderLayout
+      <Layouts.Header
         title={formatMessage({ id: 'config-sync.Header.Title' })}
         subtitle={formatMessage({ id: 'config-sync.Header.Description' })}
         as="h2"
