@@ -25,7 +25,7 @@ module.exports = {
 
 
     ctx.send({
-      message: `Config was successfully exported to ${strapi.config.get('plugin.config-sync.syncDir')}.`,
+      message: `Config was successfully exported to ${strapi.config.get('plugin::config-sync.syncDir')}.`,
     });
   },
 
@@ -37,7 +37,7 @@ module.exports = {
    */
   importAll: async (ctx) => {
     // Check for existance of the config file sync dir.
-    if (!fs.existsSync(strapi.config.get('plugin.config-sync.syncDir'))) {
+    if (!fs.existsSync(strapi.config.get('plugin::config-sync.syncDir'))) {
       ctx.send({
         message: 'No config files were found.',
       });
@@ -73,7 +73,7 @@ module.exports = {
    */
   getDiff: async (ctx) => {
     // Check for existance of the config file sync dir.
-    if (!fs.existsSync(strapi.config.get('plugin.config-sync.syncDir'))) {
+    if (!fs.existsSync(strapi.config.get('plugin::config-sync.syncDir'))) {
       ctx.send({
         message: 'No config files were found.',
       });
@@ -104,7 +104,7 @@ module.exports = {
   getAppEnv: async () => {
     return {
       env: strapi.server.app.env,
-      config: strapi.config.get('plugin.config-sync'),
+      config: strapi.config.get('plugin::config-sync'),
     };
   },
 };
